requirements.txt
fastapi
uvicorn
python-dotenv
openai
langchain-upstage
langchain-chroma
langchain-openai
chromadb
requests
tiktoken
pytz